+++ packages/eyes-sdk-core/lib/runner/EyesRunner.js
@@ -1,11 +1,14 @@
 'use strict'
 
 const {GeneralUtils} = require('@applitools/eyes-common')
+const {TestResultsSummary} = require('./TestResultsSummary')
 
 class EyesRunner {
   constructor() {
     /** @type {Eyes[]} */
     this._eyesInstances = []
+    /** @type {TestResults[]} */
+    this._allTestResult = []
     this._getBatchInfo = undefined
   }
 
@@ -26,12 +29,29 @@ class EyesRunner {
   }
 
   /**
-   * @abstract
    * @param {boolean} [throwEx=true]
    * @return {Promise<TestResultsSummary>}
    */
-  async getAllTestResults(_throwEx) {
-    throw new TypeError('The method is not implemented!')
+  async getAllTestResults(throwEx = true) {
+    await this._closeAllBatches()
+    await this._awaitAllClosePromises()
+
+    const summary = new TestResultsSummary(this._allTestResult)
+
+    if (throwEx === true) {
+      for (let result of summary.getAllResults()) {
+        if (result.getException()) {
+          throw result.getException()
+        }
+      }
+    }
+    return summary
+  }
+
+  async _awaitAllClosePromises() {
+    if (this._eyesInstances.length > 0) {
+      await Promise.all(this._eyesInstances.map(eyes => eyes._closePromise))
+    }
   }
 
   /**