+++ packages/eyes-sdk-core/lib/runner/VisualGridRunner.js
@@ -1,5 +1,6 @@
 'use strict'
-const {TestResultsSummary} = require('./TestResultsSummary')
+
+const {GeneralUtils} = require('@applitools/eyes-common')
 const {EyesRunner} = require('./EyesRunner')
 
 class VisualGridRunner extends EyesRunner {
@@ -18,34 +19,20 @@ class VisualGridRunner extends EyesRunner {
     return this._concurrentSessions
   }
 
-  /**
-   * @param {boolean} [throwEx=true]
-   * @return {Promise<TestResultsSummary>}
-   */
-  async getAllTestResults(throwEx = true) {
-    await this._closeAllBatches()
-    if (this._eyesInstances.length === 1) {
-      const [eyes] = this._eyesInstances
-      return eyes.closeAndReturnResults(throwEx)
-    } else if (this._eyesInstances.length > 1) {
-      const results = await Promise.all(
-        this._eyesInstances.map(async eyes => eyes.closeAndReturnResults(false)),
-      )
-
-      const allResults = []
-      for (const result of results) {
-        allResults.push(...result.getAllResults())
-      }
-
-      if (throwEx === true) {
-        for (const result of allResults) {
-          if (result.getException()) throw result.getException()
-        }
-      }
-      return new TestResultsSummary(allResults)
+  makeGetVisualGridClient(makeVisualGridClient) {
+    if (!this._getVisualGridClient) {
+      this._getVisualGridClient = GeneralUtils.cachify(makeVisualGridClient)
     }
+  }
 
-    return null
+  async getVisualGridClientWithCache(config) {
+    if (this._getVisualGridClient) {
+      return this._getVisualGridClient(config)
+    } else {
+      throw new Error(
+        'VisualGrid runner could not get visual grid client since makeGetVisualGridClient was not called before',
+      )
+    }
   }
 }