+++ packages/eyes-sdk-core/test/unit/server/RunningSession.spec.js
@@ -8,15 +8,15 @@ describe('RunningSession', () => {
   it('constructor without arguments', () => {
     const rs = new RunningSession()
 
-    assert.strictEqual(rs.getIsNewSession(), false)
+    assert.strictEqual(rs.getIsNew(), undefined)
     assert.strictEqual(rs.getId(), undefined)
     assert.strictEqual(rs.getUrl(), undefined)
 
-    rs.setNewSession(true)
+    rs.setIsNew(true)
     rs.setUrl('dummy url')
     rs.setId('dummy id')
 
-    assert.strictEqual(rs.getIsNewSession(), true)
+    assert.strictEqual(rs.getIsNew(), true)
     assert.strictEqual(rs.getId(), 'dummy id')
     assert.strictEqual(rs.getUrl(), 'dummy url')
   })