+++ packages/eyes-selenium/lib/EyesSelenium.js
@@ -83,6 +83,8 @@ class EyesSelenium extends Eyes {
     this._effectiveViewport = Region.EMPTY
     /** @type {EyesWebDriverScreenshotFactory} */
     this._screenshotFactory = undefined
+    /** @type {Promise<void>} */
+    this._closePromise = Promise.resolve()
   }
 
   /**
@@ -799,16 +801,24 @@ class EyesSelenium extends Eyes {
    * @return {Promise<TestResults>}
    */
   async close(throwEx = true) {
-    const results = await super.close(true).catch(err => {
-      if (!throwEx) return err
-      else throw err
-    })
+    let isErrorCaught = false
+    this._closePromise = super
+      .close(true)
+      .catch(err => {
+        isErrorCaught = true
+        return err
+      })
+      .then(results => {
+        if (this._runner) {
+          this._runner._allTestResult.push(results)
+        }
+        if (throwEx && isErrorCaught) {
+          throw results
+        }
+        return results
+      })
 
-    if (this._runner) {
-      this._runner._allTestResult.push(results)
-    }
-
-    return results
+    return this._closePromise
   }
 
   /**