+++ packages/eyes-storybook/src/defaultConfig.js
@@ -10,6 +10,7 @@ module.exports = {
   showStorybookOutput: false,
   waitBeforeScreenshot: 50,
   waitBeforeScreenshots: 50, // backward compatibility
+  viewportSize: {width: 1024, height: 768},
   tapFilePath: undefined,
   exitcode: true,
   readStoriesTimeout: 60000,