+++ packages/eyes-testcafe/lib/EyesTestCafe.js
@@ -84,6 +84,8 @@ class EyesTestCafe extends Eyes {
     this._effectiveViewport = Region.EMPTY
     /** @type {EyesWebDriverScreenshotFactory} */
     this._screenshotFactory = undefined
+    /** @type {Promise<void>} */
+    this._closePromise = Promise.resolve()
   }
 
   /**
@@ -798,16 +800,24 @@ class EyesTestCafe extends Eyes {
    * @return {Promise<TestResults>}
    */
   async close(throwEx = true) {
-    const results = await Eyes.prototype.close.call(this, true).catch(err => {
-      if (!throwEx) return err
-      else throw err
-    })
+    let isErrorCaught = false
+    this._closePromise = Eyes.prototype.close
+      .call(this, true)
+      .catch(err => {
+        isErrorCaught = true
+        return err
+      })
+      .then(results => {
+        if (this._runner) {
+          this._runner._allTestResult.push(results)
+        }
+        if (throwEx && isErrorCaught) {
+          throw results
+        }
+        return results
+      })
 
-    if (this._runner) {
-      this._runner._allTestResult.push(results)
-    }
-
-    return results
+    return this._closePromise
   }
 
   /**