+++ packages/dom-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 4.7.8
+
+- update @applitools/eyes-common@3.20.1
+
 ## 4.7.7
 
 - update @applitools/eyes-common@3.19.0