+++ packages/eyes-common/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 3.20.1
+
+- returning batchInfo from toOpenEyesConfiguration
+
 ## 3.20.0
 
 - Region supports toPersistedRegions() for VGC