+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,4 +2,20 @@
 
 ## Unreleased
 
+
+## 3.9.17
+
+- Typescript type defs: mark `cy.eyesCheckWindow` arguments as optional
+
+## 3.9.16
+
+- Support both new and old server versions for identifying new running sessions. ([Trello](https://trello.com/c/mtSiheZ9/267-support-startsession-as-long-running-task))
+
+## 3.9.15
+
+- add missing browsers to TypeScript definitions [Trello](https://trello.com/c/2QEofssH/264-cypress-js-vs-ts-not-the-same-goodrx-poc)
+
+## 3.9.14
+
+- fix trying to fetch branch info from server on non github integration runs
 - support future long running tasks [Trello](https://trello.com/c/60Rm4xXG/240-support-future-long-running-tasks)