+++ packages/eyes-testcafe/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+
+## 2.0.0-beta.17
+
+- fix version name, same version as 2.0.0-beta.16
+
+## 2.0.0-beta.16
+
+- fix trying to fetch branch info from server on non github integration runs
 - upload domsnapshot directly to Azure [Trello](https://trello.com/c/ZCLJo8Fy/241-upload-dom-directly-to-azure)
 - support future long running tasks [Trello](https://trello.com/c/60Rm4xXG/240-support-future-long-running-tasks)
 - fix regression in css stitching [Trello](https://trello.com/c/dp5IIoFw/235-css-stitching-regression-in-41533)